chore: refactor app render
