+++ stomata-cli/src/renders/mod.rs
@@ -1,3 +1,4 @@
+pub mod render_app;
 pub mod render_bar;
 pub mod render_gauge;
 pub mod render_paragraph;
